GO Game
The game of Go has established itself as the current standard by which to assess the intelligence of a computer programme surplanting chess. This is due to the fact that brute-force approaches to searching the board for optimal solutions take too long for Go. This assignment is split into separate deliverables, building up complexity with each one.
The Game:
Go is a game played on a grid. The grid we will use for this assignment is an 11x11 board. Players take turns placing stones on the board at specific grid positions, one player plays white stones, the other black. Only one stone may be placed at a time and once placed its position may not be changed. Players may choose to skip their turn. The game ends when black followed by white skip a go.


Stones may be captured. If they are, then they are removed from the board and kept by the capturer as prisoners. The winner is decided based on amount of territory and prisoners captured.
Scoring:
Each empty space surrounded by one colour is counted toward the score along with each prisoner. Each counts as one point.
In the image to the right, a prisoner was taken by black at point a. Counting the spaces surrounded by black, the top area has 5 captured spaces, the bottom has 9, the space vacated by a is 1, so that is a count of 15 then we add 1 for the captured piece so black’s total is 16. White on the other hand has 11 captured spaces bottom left, and 6 spaces top right, giving it a total of 17 and hence winning the game by 1 point. 
Rules:
The following are the rules of the game that we shall use in our implementation
Capturing Stones & Counting Liberties
The cardinal directions around a stone are called its liberties. To capture a piece, all of these spaces must be occupied. A solid connection of pieces of the one colour that are adjacent vertically and/or horizontally (not diagonally) are collectively called a string. To capture a full string, all of its liberties must be captured (that is all pieces above, beside and below the string must be filled by the other colour).


Self Capture
Self capture is not allowed (placing a piece in a position where it is immediately captured) unless it would result in capturing the surrounding pieces.
For example, we can place a white piece at i to capture the black string that surrounds it only because that black string is surrounded by white on all liberties.
We could place a white piece at j to caputure the black string around it only because the black string’s other liberties are filled by white. Without those surrounding white pieces we could not place a white at i or j. Note that we also cannot place a black piece at either of these locations.



Eyes
A string that has two separated spaces, similar to i and j above (m and n to the left) cannot be captured. To capture it would require placing a piece at both m and n simultaneously. Any string or group of stones which has two or more eyes is safe from capture.
KO Rule
You may not play a position that recaptures a position that has just been captured from you. Allowing this could result in an infinite loop.
All of these rules are fully explained at http://www.britgo.org/intro/intro2.html There are other rules at this address, however, you only need to consider those specified above.